Refactor and decompose the writer into multiple files
